Fixed spacing issue in heavily nested items.
Lists contained in dictionaries contained in lists added a great big space between the dictionary key and the list itself, when it is, in fact, supposed to be bumped right up against it. Neglected to put in some kind of mechanism to completely erase any kind of indentation or tabbing for the first item in a list.